requirements.txt
azure-functions==1.23.0
openai==1.11.0
